+++ package.json
@@ -9,7 +9,10 @@
     "website/*",
     "plugins/*"
   ],
-  "nohoist": ["website/docs/**", "**/html-minifier-terser"],
+  "nohoist": [
+    "website/docs/**",
+    "**/html-minifier-terser"
+  ],
   "description": "",
   "scripts": {
     "lint": "eslint '**/*.ts?(x)' --fix",
@@ -18,7 +21,7 @@
     "test": "jest sandpack-react",
     "cy:open": "cypress open",
     "cy:run": "cypress run",
-    "build": "yarn workspace @codesandbox/sandpack-client run build && yarn workspace @codesandbox/sandpack-react run build",
+    "build": "yarn workspace @codesandbox/sandpack-client run build && yarn workspace @codesandbox/sandpack-react run build && yarn workspace @codesandbox/sandpack-themes run build",
     "prepare": "husky install",
     "clean": "yarn workspaces run clean",
     "dev:docs": "yarn workspace sandpack-docs start",
