+++ sandpack-react/src/themes/ExternalThemes.stories.tsx
@@ -1,0 +1,11 @@
+import * as allThemes from "@codesandbox/sandpack-themes";
+import { storiesOf } from "@storybook/react";
+import React from "react";
+
+import { Sandpack } from "../";
+
+const stories = storiesOf("presets/Themes (external)", module);
+
+Object.entries(allThemes).forEach(([themeName, value]) =>
+  stories.add(themeName, () => <Sandpack theme={value} />)
+);
