+++ themes/README.md
@@ -1,0 +1,17 @@
+<img style="width:100%" src="https://user-images.githubusercontent.com/4838076/143581035-ebee5ba2-9cb1-4fe8-a05b-2f44bd69bb4b.gif" alt="Component toolkit for live running code editing experiences" />
+
+# Sandpack Themes
+
+A list of themes to customize your Sandpack components.
+
+```jsx
+import { githubLightTheme } from "@codesandbox/sandpack-themes";
+
+<Sandpack theme={githubLightTheme} />;
+```
+
+[Read more](https://sandpack.codesandbox.io/)
+
+## Documentation
+
+For full documentation, visit [https://sandpack.codesandbox.io/docs/](https://sandpack.codesandbox.io/docs/)
