+++ themes/src/githubLightTheme.ts
@@ -1,0 +1,37 @@
+import type { SandpackTheme } from "./types";
+
+export const githubLightTheme: SandpackTheme = {
+  palette: {
+    activeText: "#24292e",
+    defaultText: "#959da5",
+    inactiveText: "#e4e7eb",
+    activeBackground: "#e4e7eb",
+    defaultBackground: "#ffffff",
+    inputBackground: "#ffffff",
+    accent: "#c8c8fa",
+    errorBackground: "#ffcdca",
+    errorForeground: "#811e18",
+  },
+  syntax: {
+    keyword: "#d73a49",
+    property: "#005cc5",
+    plain: "#24292e",
+    static: "#032f62",
+    string: "#032f62",
+    definition: "#6f42c1",
+    punctuation: "#24292e",
+    tag: "#22863a",
+    comment: {
+      color: "#6a737d",
+      fontStyle: "normal",
+    },
+  },
+  typography: {
+    bodyFont:
+      '-apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol"',
+    monoFont:
+      '"Fira Mono", "DejaVu Sans Mono", Menlo, Consolas, "Liberation Mono", Monaco, "Lucida Console", monospace',
+    fontSize: "14px",
+    lineHeight: "1.4",
+  },
+};
